requirements.txt
flask==2.3.3
flask-cors==3.0.10
streamlit==1.25.0
chromadb==0.3.29
sentence-transformers==2.2.2
pdfplumber==0.10.0
pandas==2.2.2
openai==0.28.0
python-dotenv==1.0.0
tqdm==4.66.1
nltk==3.8.1
requests==2.31.0
scikit-learn==1.3.2
huggingface_hub==0.19.4
duckduckgo_search==2.7.0
beautifulsoup4==4.12.2